+++ src/main/java/com/zb/jogakjogak/security/entity/Member.java
@@ -5,13 +5,11 @@ import com.zb.jogakjogak.notification.entity.Notification;
 import com.zb.jogakjogak.resume.entity.Resume;
 import com.zb.jogakjogak.security.Role;
 import com.zb.jogakjogak.security.config.EmailEncryptor;
-import com.zb.jogakjogak.security.config.PhoneNumberEncryptor;
 import com.zb.jogakjogak.security.dto.OAuth2ResponseDto;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
-import org.springframework.context.annotation.Configuration;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -42,7 +40,7 @@ public class Member {
 
     private String nickname;
 
-    @Convert(converter = PhoneNumberEncryptor.class)
+    //@Convert(converter = PhoneNumberEncryptor.class)
     private String phoneNumber;
 
     @Enumerated(EnumType.STRING)
@@ -79,13 +77,10 @@ public class Member {
     public void updateExistingMember(OAuth2ResponseDto oAuth2ResponseDto) {
         this.email = oAuth2ResponseDto.getEmail();
         this.nickname = oAuth2ResponseDto.getNickname();
-        this.phoneNumber = oAuth2ResponseDto.getPhoneNumber();
-        this.name = oAuth2ResponseDto.getName();
     }
 
     public void setResume(Resume resume) {
         this.resume = resume;
-
         if (resume != null && (resume.getMember() == null || !resume.getMember().equals(this))) {
             resume.setMember(this);
         }
