+++ src/main/resources/db/migration/main/V4_modify_phone_number_null_able.sql
@@ -1,0 +1,1 @@
+ALTER TABLE member MODIFY phone_number VARCHAR(255) NULL;
